fast version (>5k ops/ms)

+++ src/main/java/de/creativecouple/validation/isbn/Range.java
@@ -3,15 +3,15 @@ package de.creativecouple.validation.isbn;
 class Range {
 
     final int width;
-    final int lower;
-    final int upper;
+    final long lower;
+    final long upper;
     final Range[][] rules;
 
-    Range(int width, int lower, int upper) {
+    Range(int width, long lower, long upper) {
         this(width, lower, upper, (Range[][]) null);
     }
 
-    Range(int width, int lower, int upper, Range[]... rules) {
+    Range(int width, long lower, long upper, Range[]... rules) {
         this.width = width;
         this.lower = lower;
         this.upper = upper;
